Breaking change: Update function signatures
Change all function signatures from:

  XVar(p *bool, name string, def bool, aliases ...string)

To:

  XVar(p *bool, name string, def bool, fns ...ModifyFn)

This change allows to pass different fuctions to the option that will
modify single option behaviour and will allow for multiple features
without future breaking changes in the fuction signature.

As part as this change, a new function, `opt.Alias` is added to support
previous functionality.

To update, change the aliases from a list of aliases as the variadic
last argument to a list of aliases passed to the `opt.Alias` function.
For example:

  - opt.BoolVar(&flag, "flag", false, "f", "alias2")
  + opt.BoolVar(&flag, "flag", false, opt.Alias("f", "alias2"))