Feature: Add opt.GetEnv
Requires duplicating the initialization code because we are iterating
over the modifying functions before the pointer is defined in the XxxVar
kind of functions.

It enables any modifying function to set values.

Currently only opt.String and opt.StringVar are supported, all other
option types are not supported and respond with a NO-OP.